Refactor: Move model defaults helper functions from llm_args to llm_utils
