+++ tensorrt_llm/_torch/pyexecutor/model_loader.py
@@ -9,9 +9,9 @@ import torch
 from tensorrt_llm._torch.models.checkpoints.base_checkpoint_loader import (
     AutoCheckpointMapper, BaseCheckpointLoader)
 from tensorrt_llm._utils import str_dtype_to_torch
-from tensorrt_llm.llmapi.llm_args import (TorchLlmArgs,
-                                          apply_model_defaults_to_llm_args,
-                                          validate_model_defaults)
+from tensorrt_llm.llmapi.llm_args import TorchLlmArgs
+from tensorrt_llm.llmapi.llm_utils import (apply_model_defaults_to_llm_args,
+                                           validate_model_defaults)
 from tensorrt_llm.logger import logger
 from tensorrt_llm.lora_helper import LoraConfig
 from tensorrt_llm.mapping import Mapping
