+++ tensorrt_llm/llmapi/llm_utils.py
@@ -6,11 +6,11 @@ import time
 import weakref
 from dataclasses import asdict, dataclass, field, is_dataclass
 from pathlib import Path
-from typing import Any, Callable, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import torch
 import transformers
-from pydantic import BaseModel
+from pydantic import BaseModel, ValidationError
 from tqdm import tqdm
 
 from .._utils import (global_mpi_rank, local_mpi_rank, mpi_barrier,
@@ -23,6 +23,7 @@ from ..bindings.executor import (BatchingType, CapacitySchedulerPolicy,
 from ..builder import BuildConfig, Engine, build
 from ..llmapi.llm_args import TrtLlmArgs
 from ..logger import logger
+from ..lora_helper import LoraConfig
 from ..mapping import Mapping
 from ..models.automodel import MODEL_MAP, AutoConfig, AutoModelForCausalLM
 from ..models.modeling_utils import PretrainedConfig, QuantAlgo, QuantConfig
@@ -30,16 +31,15 @@ from ..module import Module
 from .build_cache import (BuildCache, BuildCacheConfig, CachedStage,
                           get_build_cache_config_from_env)
 # yapf: disable
-from .llm_args import (CalibConfig, CudaGraphConfig, DraftTargetDecodingConfig,
+from .llm_args import (BaseLlmArgs, CalibConfig, CudaGraphConfig,
+                       DecodingConfig, DraftTargetDecodingConfig,
                        Eagle3DecodingConfig, EagleDecodingConfig, KvCacheConfig,
                        LlmArgs, LookaheadDecodingConfig, MedusaDecodingConfig,
-                       MTPDecodingConfig, NGramDecodingConfig,
-                       UserProvidedDecodingConfig, _ModelFormatKind,
-                       _ModelWrapper, _ParallelConfig,
-                       apply_model_defaults_to_llm_args,
+                       MoeConfig, MTPDecodingConfig, NGramDecodingConfig,
+                       SchedulerConfig, UserProvidedDecodingConfig,
+                       _ModelFormatKind, _ModelWrapper, _ParallelConfig,
                        update_llm_args_with_extra_dict,
-                       update_llm_args_with_extra_options,
-                       validate_model_defaults)
+                       update_llm_args_with_extra_options)
 # yapf: enable
 from .mpi_session import MPINodeState, MpiSession
 from .tokenizer import TransformersTokenizer, load_hf_tokenizer
@@ -1004,4 +1004,178 @@ __all__ = [
     'Eagle3DecodingConfig',
     'update_llm_args_with_extra_dict',
     'update_llm_args_with_extra_options',
+    'extract_llm_args_overrides',
+    'validate_model_defaults',
+    'apply_model_defaults_to_llm_args',
+    'merge_llm_configs_with_defaults',
+    'compute_applied_llm_defaults',
 ]
+
+
+# Model defaults helper functions (moved from llm_args.py)
+def extract_llm_args_overrides(llm_args) -> Dict:
+    """Extract user-provided overrides from a Pydantic model or dict."""
+    if isinstance(llm_args, dict):
+        return llm_args
+
+    fields_set = getattr(llm_args, "model_fields_set", None)
+    if fields_set is None:
+        return {}
+
+    overrides = {}
+    for field in fields_set:
+        value = getattr(llm_args, field)
+        if hasattr(value, "model_fields_set"):
+            nested = extract_llm_args_overrides(value)
+            overrides[field] = nested
+        else:
+            overrides[field] = value
+    return overrides
+
+
+def validate_model_defaults(defaults_dict: Dict, llm_args: BaseLlmArgs) -> Dict:
+    """Validate model defaults against LlmArgs schema using setattr.
+
+    Uses Pydantic's built-in validation when setting attributes.
+    Returns the validated defaults dict.
+
+    Args:
+        defaults_dict: Dictionary of model-specific defaults to validate
+        llm_args: The LlmArgs instance to validate against
+
+    Returns:
+        The validated defaults dictionary
+
+    Raises:
+        ValueError: If any field name is invalid or value has wrong type
+    """
+    if not defaults_dict:
+        return defaults_dict
+
+    # Create a temporary copy to test validation
+    temp_args = llm_args.model_copy(deep=True)
+
+    for key, value in defaults_dict.items():
+        try:
+            # This triggers Pydantic validation!
+            setattr(temp_args, key, value)
+        except (ValidationError, AttributeError) as e:
+            raise ValueError(
+                f"Invalid model default field '{key}' with value {value}. Error: {e}"
+            ) from e
+
+    return defaults_dict
+
+
+def apply_model_defaults_to_llm_args(llm_args, model_defaults: Dict) -> Dict:
+    """Apply model defaults to a Pydantic LlmArgs instance.
+
+    Returns the defaults that were actually applied.
+    """
+    if getattr(llm_args, "_model_defaults_applied", False):
+        return {}
+    if not model_defaults:
+        return {}
+
+    llm_args_dict = llm_args.model_dump()
+    user_overrides = extract_llm_args_overrides(llm_args)
+    merged_args = merge_llm_configs_with_defaults(llm_args_dict,
+                                                  model_defaults,
+                                                  user_overrides=user_overrides)
+
+    for key in model_defaults:
+        if key in merged_args and hasattr(llm_args, key):
+            setattr(llm_args, key, merged_args[key])
+
+    applied = compute_applied_llm_defaults(model_defaults, user_overrides)
+    llm_args._model_defaults_applied = True
+    return applied
+
+
+def merge_llm_configs_with_defaults(
+    llm_args: Dict,
+    model_defaults: Dict,
+    user_overrides: Optional[Dict] = None,
+) -> Dict:
+    """Merge model-specific defaults into llm_args using Pydantic.
+
+    Only applies defaults for fields that are unset in llm_args.
+    Uses Pydantic's model_validate and model_dump(exclude_unset=True).
+    """
+    # Handle nested Pydantic models generically
+    config_classes = {
+        "kv_cache_config": KvCacheConfig,
+        "build_config": BuildConfig,
+        "scheduler_config": SchedulerConfig,
+        "decoding_config": DecodingConfig,
+        "quant_config": QuantConfig,
+        "lora_config": LoraConfig,
+        "moe_config": MoeConfig,
+    }
+
+    if user_overrides is None:
+        user_overrides = llm_args if isinstance(llm_args, dict) else {}
+    else:
+        user_overrides = user_overrides or {}
+    for key, default_value in model_defaults.items():
+        if key in config_classes:
+            config_class = config_classes[key]
+
+            # Get existing config or empty dict
+            existing = llm_args.get(key, {})
+            if hasattr(existing, 'model_dump'):
+                # If it's already a Pydantic model, convert to dict
+                existing = existing.model_dump()
+            elif not isinstance(existing, dict):
+                # If it's not a dict, skip
+                continue
+
+            override_value = user_overrides.get(key, {})
+            if hasattr(override_value, "model_dump"):
+                override_value = override_value.model_dump(exclude_unset=True)
+            if not isinstance(override_value, dict):
+                override_value = {}
+
+            merged = existing.copy()
+            for sub_key, sub_default in default_value.items():
+                if sub_key not in override_value:
+                    merged[sub_key] = sub_default
+            llm_args[key] = config_class.model_validate(merged)
+        else:
+            # Simple fields - only set if not already present
+            if key not in user_overrides:
+                llm_args[key] = default_value
+
+    return llm_args
+
+
+def compute_applied_llm_defaults(model_defaults: Dict,
+                                 user_overrides: Dict) -> Dict:
+    """Return the subset of model_defaults that will be applied.
+
+    Only defaults for fields not present in user_overrides are included.
+    Nested dicts are handled recursively.
+    """
+
+    def _compute(defaults: Dict, overrides: Dict) -> Dict:
+        applied = {}
+        overrides = overrides or {}
+        for key, default_value in defaults.items():
+            if isinstance(default_value, dict):
+                override_value = overrides.get(key) if isinstance(
+                    overrides, dict) else None
+                if hasattr(override_value, "model_dump"):
+                    override_value = override_value.model_dump(
+                        exclude_unset=True)
+                if isinstance(override_value, dict):
+                    nested = _compute(default_value, override_value)
+                    if nested:
+                        applied[key] = nested
+                else:
+                    applied[key] = default_value
+            else:
+                if not isinstance(overrides, dict) or key not in overrides:
+                    applied[key] = default_value
+        return applied
+
+    return _compute(model_defaults, user_overrides or {})
